fix: CORS to main.rs
